Always try to show the re-rolled cat
Previously, I decided to not show this because it's impossible to
reach it from there so it's impossible to have dupe when the
last one is re-rolled. But it can be re-rolled from the other
track! When that happen, the following rare cat is still reachable
therefore it can still be a dupe.

I can't seem to find this edge case, but in theory it can happen.
To also account this, we need to back track to the very beginning!
In theory it's possible to 1A -> 2B -> 4A -> 5B -> 7A -> ...
But in practice of course it's unlikely to happen, not to mention
we're not using 64 bits PRNG.

It's just easier to always try to show the information, considering
that people can reach it with platinum or guaranteed anyway...
Better to show more information than missing information!