Remove the hacky next_seed code
We can do this by advancing the seed before trying, so surely we know
what was the previous seed. The problem is that, now we're no longer
looking at all the possible numbers, but the offset. For example,
suppose we're looking at 0, 1, 2, 3, 4. Now we're looking at
offset(0), offset(1), offset(2), offset(3), offset(4). It can be
possible that an initial seed number will never loop back and
we won't check them.

However, this is unlikely to be an issue anyway. Because they
shouldn't really use any initial seed number anyway... They
should always rotate once like we did!