Fixed the edge case for missing guaranteed
Fixed an edge case where guaranteed cats might be missing for some cells.
It was happening when one of the cells had a dupe from both A and B, and
when that happened next cat data would be lost causing missing guaranteed.
It was only discovered with Merc Storia because it has the most dupes.