Use the regular cats map if possible
Previously, the very original data is indexed by (rarity -> id -> cat).
This is useful in some cases, but can be difficult to use if we only
want to look up by id. This was intended because:

* I wanted to avoid embedding the rarity data into the cat data
* It's cheaper to transform this, than the other way around

However, with the introduction of showing cats stat, this is getting
more and more difficult to use because most of the time we want to
traverse through all the cats, regardless rarity. Flatten the raw data
into simply (id -> cat with rarity) is simpler and easier to understand.

If we already want to embed the data already, why not we reverse the
main map, too? Clearly in a lot of cases this is easier to use, too.

So in the end not much difference in terms of performance, but this
should make the code and data easier to use and follow.