Change discount DPS option to only for summation
It's odd that if we discount DPS for wave then wave has 0 DPS.
It's also odd that even if this option is used, damage sum still
counts wave damage, and it'll be odd to make damage to be 0 with that,
too.

So ideally this option should only control the sum values, which
should include DPS and damage, and nothing else.