Allow pages not for tracking to be crawled
Tracking has infinite pages and links, and the information is not that
helpful for a bot, unless it's made to be assisting tracking.

It could be cool if AI can understand tracking, but I do not want to
exhaust the site for AI bots. The resources should be reserved for
actual users/players.