Always remember the last rolled cat
We do this by setting the id as the value in the default blank option.
This will not show the cat name, which is fine, because it doesn't
really matter if it doesn't exist in the banner we're looking at,
and in that case it'll preserve the value. When a banner has the cat,
the actual cat option will be selected and the blank option will not be
used.

This means you'll not be able to drop the value unless you edit the URL,
which is probably fine because we probably wouldn't need to drop it
anyway.

The only case we might want to drop it might be when we can retreat the
seed, which we cannot do in the game therefore the last rolled cat will
not make sense anymore. However, we cannot do this on the site yet.
Maybe we can revisit this when it's a feature. When we retreat, remove
the last cat maybe? Because it's likely it's no longer accurate anyway.