Added high effective single blow filter
This also changes so that we only look at true single blow,
and ignores damage sum. It's done this way because it doesn't
really make sense for Gaia to have very high single blow when
it's effectively divided between 3 attacks. It's the same cycle
but the delay between them is long enough that it doesn't make
sense to be considered as a "single blow".

Also, the reason why we would want to consider "single blow" is that
there are enemies that can be knocked back easily, and multi-hits
can miss easily for those cases so having a high single blow can be
useful for those enemies. Gaia doesn't really help in this case.