Update and download server data for BCKR and BCTW
So now they moved the animation data to server which is indicated by
`download_34.tsv` file specifically in `WImageDataServer.pack`.
This file doesn't indicate which version id to use, and fortunately
that's just the current version. This seems to be hard coded in the
binary .so file so it could be annoying to actually parse this.

For now, we can just preserve the file and version mapping like we
did for the gacha. Hopefully in the future they can just put this
in the download tsv files.