Add game version for SImageDataServer in 29
I figured this out by looking at the historical versions,
and tried it out one by one. Not that this is useful though,
because it's proven that there's nothing new there. However,
I guess for the sake of completeness, I am adding it.

If they're not used in the future, they should be removing
this from the game, and then we won't download this again.
For now, let's just grab all data as much as we can in case
we might be missing something.