Add a new TODO to archive past events
There's an interesting discussion that the ordering of the events isn't
consistent between upcoming and past events. For the upcoming ones,
it's chronological, but for past ones, it's reverse chronological.

I can see that can be confusing if one doesn't know if they're looking
at the upcoming section or the past section. On a computer it's
relatively clear, but on a mobile device it's not so clear because the
section label is so small and the viewport is also much smaller, making
it more difficult to see the big picture.

If we archive the past events, putting them behind another selection,
like how we did that for the customized banners, we should be able to
make this less confusing because if the visitor is looking at the past
events, they should know that because they need to pick that first,
and from there, the events would be showing in a completely different
dropdown, making the ordering consistent within the same dropdown.

This can look like this:

* Upcoming:
  * Yesterday
  * Today
  * Tomorrow
* Custom:
  * Customize...
* Past:
  * Within last year...
  * Within 2 years ago...
  * Within 3 years ago...

When clicking on "Customize..." and "Within last year..." should go to
another page, showing a separate dropdown to select the custom banners
or past events, in reverse chronological order.

This way we should be keeping the upcoming one as chronological and it's
still consistent within the same dropdown, making it less confusing.

This can also address a long standing concern of mine, that eventually
the event list must need some kind of pagination because it can grow
indefinitely until this game is dead, which we certainly hope that is,
never. This should improve performance and usability when it's really
big.