Avoid crashing the stat page when names and descriptions aren't aligned
This was crashing: https://bc.godfat.org/cats/789?lang=kr
Because there are 3 names but only 2 descriptions.

The Cat object handles this by going for a previous form, but
the Stat object doesn't know about the previous form. If we pass
the Cat object to the Stat object, then we can use the method for
this particular fallback.

This feels a bit messy because why should we have multiple ways to
access the same information? We could potentially always access that
from the Cat object? But then it's less flexible because in the case of
filtering, we don't even have a Cat object and we definitely would
already have the info hash we can access with anyway. Why should we
make this more coupled after all?

So this messy way to try to access all data if possible might be the
better compromise after all. Not sure, at least the code is still
relatively simple anyway I guess.