Fix seed seeker not getting the right user input
This was broken by https://github.com/rack/rack/pull/2351/files

Changelog for this was:
https://github.com/rack/rack/blob/v3.2.0/CHANGELOG.md?plain=1#L22

> - Add `Rack::Request#form_pairs` to access form data as raw key-value pairs,
>   preserving duplicate keys. ([#2351](https://github.com/rack/rack/pull/2351),
>   [@matthewd](https://github.com/matthewd))

Previously, we use `parse_query` for `POST` when it's form data:
https://github.com/rack/rack/blob/v3.1.16/lib/rack/request.rb#L538

We overrode that to provide duplicated queries:
https://gitlab.com/godfat/battle-cats-rolls/-/blob/7dc5c0c5a97d969aaa8057fda399c0773f887be7/lib/battle-cats-rolls/request.rb#L7-9

However, after the change, it always calls `expand_param_pairs` to remove
duplicates for `POST`:
https://github.com/rack/rack/blob/v3.2.0/lib/rack/request.rb#L547 and
https://github.com/rack/rack/pull/2351/files#diff-7ce97931f18a63a4d028696a6f4ba81991644dbc2d70eaa664285264e9a5cd64R547

This means we can no longer access the list of rolls. It always only gives the
one in the last slot (earlier queries are all removed)

This, to be honest, while it's a change, it's more consistent because for
multipart it's always going through this even in earlier versions:
https://github.com/rack/rack/blob/v3.1.16/lib/rack/request.rb#L529

It's an unfortunate internal change breaking how we're overriding `parse_query`
to provide duplicated queries.

In order to fix this, we now also override `expand_param_pairs` to ensure that
it doesn't call `normalize_params` to remove duplicates. This reuses part of
the original code, and also `parse_query` to construct the end result.